Update README.md
My work as an independent researcher is quite a new under-taking. I wanted to introduce this idea to all like minded professionals.  I know that majority of the work had been done by Masakhane. I am humbly urging all to at least look into this and see how we can work together to further develop it. The Congo-Kwa languages have some similarities in deriving their meaning from the same word if suffix and prefix vowels are isolated leaving the root: Meaning could be found with understanding the connections of the word and reality. In this case the root could be bi-gram,  tri-gram or sub-words of nth vocabularies. Exploring this will give African languages an edge in the age of AI.  We can move faster on our own without others who do not share the same opportunity cost.  I propose the vowel suffix or prefix to root linking to further the study and swiftly arrive to GPT.  using our ancient knowledge of IFA/AFA we can expand the language's predictive powers. -This will clearly show that language is the heartbeat of creativity.